Bug fixes gravitycorr + option to use gyro temp

+++ src/gyro.h
@@ -67,7 +67,7 @@ class GyroSensor {
         void calibrateSensor();
 
         double getAngle() { return angle; };
-        double getSensorTempC() { return sensorTemp; };
+        float  getSensorTempC() { return sensorTemp; };
         bool   isConnected() { return sensorConnected; };
         bool   hasValue() { return validValue; };
         void   enterSleep();
